Finish probdist. Change to little-endian in both languages. Fix bug with continue in graph resolver. More things changed to floats. Fix debug message in GraphDecoder.pl (empty array)
The two implementations now work more or less the same but there
are some differences:

* C version seems to solve things in a different order, sometimes
  leading to extra aux blocks being solved
* for mindecoder -d 7, C version solves on the 9th check block
  while the Perl version needs 11 check blocks
* for mindecoder -d 10, C version runs out of node space. The
  check block generation routine is coming up with differences

+++ Net-OnlineCode/C/rng_sha1.c
@@ -140,7 +140,7 @@ int oc_rng_random_uuid(char *dest) {
 
 }
 
-float oc_rng_rand(oc_rng_sha1 *rng, float max) {
+double oc_rng_rand(oc_rng_sha1 *rng, double max) {
 
   double  temp;			    // prevent overflow
   uint32_t sha_word;
